add to street segment sidebar
also move to lib/entity
